# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SunSide API — sun-side seating planner for commuters
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface GeocodeSuggestion {
  /** Human-readable address */
  displayName: string;
  /** Latitude */
  lat: number;
  /** Longitude */
  lon: number;
  /** Unique place identifier */
  placeId: string;
  /** Distance in km from the reference point (user location or map center), if available */
  distanceKm?: number;
}

export type VehicleType = typeof VehicleType[keyof typeof VehicleType];


export const VehicleType = {
  Bus: 'Bus',
  Microbus: 'Microbus',
  Car: 'Car',
  Train: 'Train',
  Metro: 'Metro',
} as const;

export interface TripLegInput {
  vehicleType: VehicleType;
  /** Human-readable origin label */
  fromLabel: string;
  /** Human-readable destination label */
  toLabel: string;
  fromLat: number;
  fromLon: number;
  toLat: number;
  toLon: number;
  /** ISO 8601 datetime string for departure */
  departureTime: string;
}

export interface TripInput {
  /** @minItems 1 */
  legs: TripLegInput[];
}

/**
 * Temperature comfort badge
 */
export type WeatherInfoBadge = typeof WeatherInfoBadge[keyof typeof WeatherInfoBadge];


export const WeatherInfoBadge = {
  Hot: 'Hot',
  Warm: 'Warm',
  Cool: 'Cool',
} as const;

export interface WeatherInfo {
  /** Whether weather data was successfully fetched */
  available: boolean;
  /** Temperature in Celsius */
  temperatureCelsius: number;
  /** WMO weather code */
  weatherCode: number;
  /** Human-readable weather description */
  description: string;
  /** Temperature comfort badge */
  badge: WeatherInfoBadge;
  /** ISO 8601 timestamp of the matched weather reading (local time) */
  weatherTimestamp?: string;
  /** Reason weather is unavailable, when available is false */
  unavailableReason?: string;
}

export type SunSampleSide = typeof SunSampleSide[keyof typeof SunSampleSide];


export const SunSampleSide = {
  Left: 'Left',
  Right: 'Right',
  None: 'None',
} as const;

export interface SunSample {
  minutesElapsed: number;
  azimuthDeg: number;
  altitudeDeg: number;
  side: SunSampleSide;
}

/**
 * Which side of the vehicle the sun is on during this segment
 */
export type SegmentBreakdownItemSunSide = typeof SegmentBreakdownItemSunSide[keyof typeof SegmentBreakdownItemSunSide];


export const SegmentBreakdownItemSunSide = {
  Left: 'Left',
  Right: 'Right',
  None: 'None',
} as const;

export interface SegmentBreakdownItem {
  /** Minutes elapsed from departure when this segment begins */
  startMinute: number;
  /** Minutes elapsed from departure when this segment ends */
  endMinute: number;
  /** Duration of this segment in minutes */
  durationMinutes: number;
  /** Average compass bearing for this segment (0-360) */
  bearingDeg: number;
  /** Human-readable heading direction (e.g. "north", "east") */
  headingLabel: string;
  /** Which side of the vehicle the sun is on during this segment */
  sunSide: SegmentBreakdownItemSunSide;
}

/**
 * Which side of the vehicle faces the sun for most of the trip
 */
export type LegResultSunSide = typeof LegResultSunSide[keyof typeof LegResultSunSide];


export const LegResultSunSide = {
  Left: 'Left',
  Right: 'Right',
  None: 'None',
} as const;

/**
 * Routing data source used for this leg
 */
export type LegResultRoutingSource = typeof LegResultRoutingSource[keyof typeof LegResultRoutingSource];


export const LegResultRoutingSource = {
  osrm: 'osrm',
  openrouteservice: 'openrouteservice',
  straight_line: 'straight_line',
  straight_line_rail: 'straight_line_rail',
} as const;

export interface LegResult {
  legIndex: number;
  vehicleType: VehicleType;
  fromLabel: string;
  toLabel: string;
  /** Travel compass bearing in degrees (0-360) */
  bearing: number;
  /** Estimated trip duration in minutes */
  durationMinutes: number;
  /** Distance in kilometers */
  distanceKm: number;
  /** Which side of the vehicle faces the sun for most of the trip */
  sunSide: LegResultSunSide;
  /** Percentage of trip where sun is on the left */
  sunPercentLeft: number;
  /** Percentage of trip where sun is on the right */
  sunPercentRight: number;
  /** Human-readable seating recommendation */
  recommendation: string;
  weather: WeatherInfo;
  /** Time-sampled sun positions across the trip */
  samples: SunSample[];
  /** Routing data source used for this leg */
  routingSource: LegResultRoutingSource;
  /**
     * List of [lat, lon] pairs tracing the route path
     * @items.minItems 2
     * @items.maxItems 2
     */
  routeCoordinates: number[][];
  /** Per-segment sun-side breakdown showing how sun exposure changes as the route curves */
  segmentBreakdown: SegmentBreakdownItem[];
}

export interface TripAnalysis {
  legs: LegResult[];
  /** ISO 8601 timestamp when analysis was performed */
  analyzedAt: string;
}

export type GeocodeAddressParams = {
/**
 * Address search query
 */
q: string;
/**
 * Maximum number of results
 */
limit?: number;
/**
 * User's current latitude for proximity sorting
 */
userLat?: number;
/**
 * User's current longitude for proximity sorting
 */
userLon?: number;
};