# lib/api-zod/src/generated/types/geocodeSuggestion.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SunSide API — sun-side seating planner for commuters
 * OpenAPI spec version: 0.1.0
 */

export interface GeocodeSuggestion {
  /** Human-readable address */
  displayName: string;
  /** Latitude */
  lat: number;
  /** Longitude */
  lon: number;
  /** Unique place identifier */
  placeId: string;
  /** Distance in km from the reference point (user location or map center), if available */
  distanceKm?: number;
}