# lib/api-zod/src/generated/types/legResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SunSide API — sun-side seating planner for commuters
 * OpenAPI spec version: 0.1.0
 */
import type { LegResultRoutingSource } from './legResultRoutingSource';
import type { LegResultSunSide } from './legResultSunSide';
import type { SegmentBreakdownItem } from './segmentBreakdownItem';
import type { SunSample } from './sunSample';
import type { VehicleType } from './vehicleType';
import type { WeatherInfo } from './weatherInfo';

export interface LegResult {
  legIndex: number;
  vehicleType: VehicleType;
  fromLabel: string;
  toLabel: string;
  /** Travel compass bearing in degrees (0-360) */
  bearing: number;
  /** Estimated trip duration in minutes */
  durationMinutes: number;
  /** Distance in kilometers */
  distanceKm: number;
  /** Which side of the vehicle faces the sun for most of the trip */
  sunSide: LegResultSunSide;
  /** Percentage of trip where sun is on the left */
  sunPercentLeft: number;
  /** Percentage of trip where sun is on the right */
  sunPercentRight: number;
  /** Human-readable seating recommendation */
  recommendation: string;
  weather: WeatherInfo;
  /** Time-sampled sun positions across the trip */
  samples: SunSample[];
  /** Routing data source used for this leg */
  routingSource: LegResultRoutingSource;
  /**
     * List of [lat, lon] pairs tracing the route path
     * @items.minItems 2
     * @items.maxItems 2
     */
  routeCoordinates: number[][];
  /** Per-segment sun-side breakdown showing how sun exposure changes as the route curves */
  segmentBreakdown: SegmentBreakdownItem[];
}