# lib/api-zod/src/generated/types/segmentBreakdownItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SunSide API — sun-side seating planner for commuters
 * OpenAPI spec version: 0.1.0
 */
import type { SegmentBreakdownItemSunSide } from './segmentBreakdownItemSunSide';

export interface SegmentBreakdownItem {
  /** Minutes elapsed from departure when this segment begins */
  startMinute: number;
  /** Minutes elapsed from departure when this segment ends */
  endMinute: number;
  /** Duration of this segment in minutes */
  durationMinutes: number;
  /** Average compass bearing for this segment (0-360) */
  bearingDeg: number;
  /** Human-readable heading direction (e.g. "north", "east") */
  headingLabel: string;
  /** Which side of the vehicle the sun is on during this segment */
  sunSide: SegmentBreakdownItemSunSide;
}