# lib/api-zod/src/generated/types/sunSample.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SunSide API — sun-side seating planner for commuters
 * OpenAPI spec version: 0.1.0
 */
import type { SunSampleSide } from './sunSampleSide';

export interface SunSample {
  minutesElapsed: number;
  azimuthDeg: number;
  altitudeDeg: number;
  side: SunSampleSide;
}