# lib/api-zod/src/generated/types/tripAnalysis.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SunSide API — sun-side seating planner for commuters
 * OpenAPI spec version: 0.1.0
 */
import type { LegResult } from './legResult';

export interface TripAnalysis {
  legs: LegResult[];
  /** ISO 8601 timestamp when analysis was performed */
  analyzedAt: string;
}