# lib/api-zod/src/generated/types/tripLegInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SunSide API — sun-side seating planner for commuters
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleType } from './vehicleType';

export interface TripLegInput {
  vehicleType: VehicleType;
  /** Human-readable origin label */
  fromLabel: string;
  /** Human-readable destination label */
  toLabel: string;
  fromLat: number;
  fromLon: number;
  toLat: number;
  toLon: number;
  /** ISO 8601 datetime string for departure */
  departureTime: string;
}