# lib/api-zod/src/generated/types/weatherInfo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SunSide API — sun-side seating planner for commuters
 * OpenAPI spec version: 0.1.0
 */
import type { WeatherInfoBadge } from './weatherInfoBadge';

export interface WeatherInfo {
  /** Whether weather data was successfully fetched */
  available: boolean;
  /** Temperature in Celsius */
  temperatureCelsius: number;
  /** WMO weather code */
  weatherCode: number;
  /** Human-readable weather description */
  description: string;
  /** Temperature comfort badge */
  badge: WeatherInfoBadge;
  /** ISO 8601 timestamp of the matched weather reading (local time) */
  weatherTimestamp?: string;
  /** Reason weather is unavailable, when available is false */
  unavailableReason?: string;
}